refactor: create `background-page` folder to group `background components`
- relocate ContentArea, Timeline, and UserInfo components into src/components/custom/background-page/...
- update all affected imports
- adjusted BackgroundContext and hooks (useTimelineBlackBorder, useTimelineNavigation) to reference the new timeline utils path
- switched some UI imports to use the '@/components/ui' alias
