+++ src/app/background/page.view.tsx
@@ -3,17 +3,17 @@
 import {
   UserInfoDesktop,
   UserInfoMobile,
-} from "../../components/custom/UserInfo";
+} from "../../components/custom/background-page/UserInfo";
 import {
   Timeline,
   TimelineNavButtonsMobile,
-} from "../../components/custom/Timeline";
+} from "../../components/custom/background-page/Timeline";
 import "./styles.css";
 import { userData } from "@/background-data";
 import { BackgroundProvider } from "../../contexts/BackgroundContext";
-import { ContentArea } from "@/components/custom/ContentArea";
+import { ContentArea } from "@/components/custom/background-page/ContentArea";
 import { useMediaQuery } from "@/hooks/useMediaQuery";
-import YearTitle from "@/components/custom/ContentArea/YearTitle";
+import YearTitle from "@/components/custom/background-page/ContentArea/YearTitle";
 
 export default function BackgroundPage() {
   const isDesktop = useMediaQuery("(min-width: 769px)");
