+++ src/components/custom/background-page/ContentArea/ContentArea.tsx
