+++ src/components/custom/background-page/ContentArea/ContentItem.tsx
