+++ src/components/custom/background-page/ContentArea/LocationInfo.tsx
