+++ src/components/custom/background-page/ContentArea/MonthBullet.tsx
