+++ src/components/custom/background-page/ContentArea/PeriodInfo.tsx
