+++ src/components/custom/background-page/ContentArea/YearTitle.tsx
