+++ src/components/custom/background-page/ContentArea/colors.utils.ts
