+++ src/components/custom/background-page/ContentArea/content-item.types.ts
