+++ src/components/custom/background-page/ContentArea/content-item.utils.ts
