+++ src/components/custom/background-page/ContentArea/content-list.utils.ts
