+++ src/components/custom/background-page/ContentArea/index.tsx
