+++ src/components/custom/background-page/ContentArea/month-bullet.utils.ts
