+++ src/components/custom/background-page/UserInfo/UserExternalLinks.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react";
-import { Button } from "../../ui/button";
+import { Button } from "@/components/ui/button";
 import Link from "next/link";
 import {
   GithubLogoIcon,
